refactor: streamline Slack action context types and improve context handling utilities
