+++ server/slack/actions/retry-ping.ts
@@ -1,11 +1,6 @@
-import type {
-  AllMiddlewareArgs,
-  BlockAction,
-  ButtonAction,
-  SlackActionMiddlewareArgs,
-} from '@slack/bolt';
 import { Actions, Button, Section } from 'slack-block-builder';
 import { asBlocks } from '~/lib/slack/blocks';
+import type { SlackButtonActionContext } from '~/types';
 
 export const name = 'retry_ping';
 
@@ -13,7 +8,7 @@ export async function execute({
   ack,
   client,
   respond,
-}: SlackActionMiddlewareArgs<BlockAction<ButtonAction>> & AllMiddlewareArgs) {
+}: SlackButtonActionContext) {
   await ack();
   const start = Date.now();
   await client.auth.test();
